backend/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
celery==5.3.4
django-cors-headers==4.2.0
python-dateutil==2.8.2
django-celery-results==2.2.0